renamed core class files according to PEP 8
renamed core class files according to CapWords convention for easier distinction from non-class files
