+++ hypergraphx/core/IHypergraph.py
